Update README with Excel file format details
Added a new section to the README file that describes the Excel file format used by the RuleEngine project. Included subsections for "Sheet one (Metadata)" and "Sheet two (Rules)" with corresponding images to illustrate the format. This addition aims to help users understand the structure and content of the Excel files that need to be uploaded to the application.